Replacing all references to CGFLoat with more appropriate float types

+++ Classes/StravaActivityZone.h
@@ -25,8 +25,8 @@ typedef NS_ENUM(NSInteger, kActivityZoneType) {
 ///
 @interface StravaActivityZoneDistributionBucket : MTLModel<MTLJSONSerializing>
 
-@property (nonatomic, readonly) CGFloat min;
-@property (nonatomic, readonly) CGFloat max;
+@property (nonatomic, readonly) float min;
+@property (nonatomic, readonly) float max;
 @property (nonatomic, readonly) NSTimeInterval time;
 
 @end
@@ -43,11 +43,11 @@ typedef NS_ENUM(NSInteger, kActivityZoneType) {
 @property (nonatomic, readonly) kActivityZoneType type;
 @property (nonatomic, readonly) kResourceState resourceState;
 @property (nonatomic, readonly) BOOL sensorBased;
-@property (nonatomic, readonly) CGFloat points;
+@property (nonatomic, readonly) float points;
 @property (nonatomic, readonly) BOOL customZones;
-@property (nonatomic, readonly) CGFloat max;
-@property (nonatomic, readonly) CGFloat bikeWeight;
-@property (nonatomic, readonly) CGFloat athleteWeight;
+@property (nonatomic, readonly) float max;
+@property (nonatomic, readonly) float bikeWeight;
+@property (nonatomic, readonly) float athleteWeight;
 @property (nonatomic, readonly, copy) NSArray *distributionBuckets; /// NSArray of StravaActivityZoneDistributionBucket objects
 
 @end
